backend/agents/vision_agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
langgraph==0.0.20
langchain-core==0.1.16
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.60.0
requests==2.31.0
python-multipart==0.0.6